requirements.txt
fastapi
uvicorn
pandas
pdfplumber
python-dotenv
langchain
openai
base64io
python-multipart
openpyxl
